app/requirements.txt
fastapi
uvicorn
pandas
torch
google-cloud-storage
google-cloud-aiplatform[prediction]
